# app/swagger.yaml
openapi: 3.0.3
info:
  title: LorrySwap API - OpenAPI 3.0
  termsOfService: http://swagger.io/terms/
  contact:
    email: fernandowkd.21@itfac.mrt.ac.lk
  version: 1.0.11
servers:
  - url: https://lorryswapbackendtest2.azurewebsites.net  
...
paths:
  /user/createacc:
    post:
      tags:
        - User
      summary: Create user
      description: Register a new user with their personal information. This can only be done during user signup.
      operationId: user.create_user
      requestBody:
        required: true
        description: User registration details.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: OTP sent to user's email. Please verify to complete registration.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "OTP sent to your email. Please verify to complete registration."
        '400':
          description: Invalid input supplied.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Invalid input - Email format is incorrect."
        '500':
          description: Internal Server Error - Error during the process.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal server error. Please try again later."
  /user/verify_otp:
    post:
      tags:
        - User
      summary: Verify OTP
      description: Verify the OTP sent to the user's email to complete the registration process.
      operationId: user.verify_otp
      requestBody:
        required: true
        description: OTP verification details.
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: michelstarc@example.com
                otp:
                  type: integer
                  example: 123456
      responses:
        '200':
          description: User created successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User created successfully."
        '400':
          description: Invalid input or OTP.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Email and OTP are required."
        '409':
          description: Conflict - Username or email already exists.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Conflict - Email already exists."
        '500':
          description: Internal Server Error - Error during the process.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal server error. Please try again later."
  /user/login_user:
    post:
      tags:
        - User
      summary: User Login
      description: Authenticate user and return JWT token including roles.
      operationId: user.login_user
      requestBody:
        required: true
        description: User login details.
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: starc123
                password:
                  type: string
                  format: password
                  example: 'Password!234'
      responses:
        '200':
          description: Login successful and token returned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  message:
                    type: string
                    example: "Login Successful"
        '401':
          description: Incorrect username or password.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Incorrect username or password"
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User not found"
        '500':
          description: Database error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Database error"
                  error:
                    type: string
                    example: "Error details"
  /user/forgot_password:
    post:
      tags:
        - User
      summary: Forgot Password
      description: Initiates the forgot password process by sending an OTP to the user's email.
      operationId: user.forgot_password
      requestBody:
        required: true
        description: Email address of the user who wants to reset the password.
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: michelstarc@example.com
      responses:
        '200':
          description: OTP sent to user's email.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "OTP sent to your email. Please verify to reset your password."
        '400':
          description: Invalid input supplied.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Email is required."
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User with this email does not exist."
        '500':
          description: Internal Server Error - Error during the process.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal server error. Please try again later."
  /user/reset_password:
    post:
      tags:
        - User
      summary: Reset Password
      description: Resets the user's password after verifying the OTP sent to their email.
      operationId: user.reset_password
      requestBody:
        required: true
        description: OTP verification and new password details.
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: michelstarc@example.com
                otp:
                  type: integer
                  example: 123456
                new_password:
                  type: string
                  format: password
                  example: 'NewPassword!234'
      responses:
        '200':
          description: Password reset successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Password reset successfully."
        '400':
          description: Invalid input or OTP.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Email, OTP, and new password are required."
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User not found."
        '500':
          description: Internal Server Error - Error during the process.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Internal server error. Please try again later."                  
...
components:
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: auth.bearer_info_func
  schemas:
    User:
      type: object
      properties:
        roles:
          type: array
          items:
            type: string
            enum:
              - Driver
              - Owner
              - CargoFinder
              - EconomicCenterSeller
              - MarketSeller
              - Admin
              - SuperAdmin
          example: ['Owner', 'Driver']
        fname:
          type: string
          example: Michel
        lname:
          type: string
          example: Starc
        email:
          type: string
          format: email
          example: michelstarc@example.com
        contact_no:
          type: string
          example: '+1234567890'
        age:
          type: integer
          format: int32
          example: 30
        gender:
          type: string
          example: m
        nic:
          type: string
          example: '987654321V'
        username:
          type: string
          example: starc123
        password:
          type: string
          format: password
          example: 'Password!234'
      required:
        - roles
        - fname
        - lname
        - email
        - contact_no
        - age
        - gender
        - nic
        - username
        - password
...